Fix and refactor Dimension Control Add responsive style for frontend
Dimension Control component needs a last review for the refactoring, but it's mostly done.
The responsive styles works for both backend and frontend, and they have been made in order to be reused in other situations.
Some component are affected for Dimension Control component changes, and need to be adapted; components such: padding&margin, border radius and border width.
Image Box Extra block needs to be checked after all this changes too.