Button visibility
Fixes #2852

Rus added 0 opacity to the icons, in order to show them on hover only. See https://github.com/yeahcan/maxi-blocks/pull/2775.

However, it seems that we changed the DOM and where the replace button is rendered. It used to be a part of the module, and now it's outside of it. Because of that, the CSS I removed is not working.

Since the replace button isn't inside the block anymore, we can't make it work easily with CSS. We probably need a function to detect block hover to add a class to the popover settings. Do we want to do that or it's complicating too much?