Added default SC values
Ok, so I think I figured out what's happening. When I added "true" for the avoidSC prop, the value wasn't null anymore, because it was getting it from the SelectControl. Ofc, SC values weren't working anymore. So the issue was that we didn't have default values for our default SC for text orientation and direction so it didn't have any values to show in the first place.