issue-2937-button-text-select-error (#2974)
* Added defaults

Fixes #2937

Problem: Text direction is causing the error. It needs a default value.
I'm not sure if this is the correct way to fix it. If not, please suggest where else can I define the default.

* SelectControl update

* Duplicate attr and text-transform

* Added default SC values

Ok, so I think I figured out what's happening. When I added "true" for the avoidSC prop, the value wasn't null anymore, because it was getting it from the SelectControl. Ofc, SC values weren't working anymore. So the issue was that we didn't have default values for our default SC for text orientation and direction so it didn't have any values to show in the first place.

* Update paragraph.spec.js.snap

* Updated tests and snapshots

* Fixes tests

Co-authored-by: Josh <jjuancap@gmail.com>